Implement new customer banner ad
Implement a banner ad at the top of the page that displays a unique, single-use discount code for new customers. The banner includes a 24-hour countdown timer, and the code can be easily copied to the clipboard. The banner is removed after 24 hours, upon successful checkout, or if the user has already used the code.